refactor: init app state
